Moved the home view from users to main
