+++ editdojo_project/urls.py
@@ -15,9 +15,12 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
+from users.views import select_languages, signup_flow
+from main.views import home
+
+# The following two lines are for CS Dojo's tutorials
 from hello.views import my_view, home_view
 from todo.views import todo_view, add_todo, delete_todo
-from users.views import home, select_languages, signup_flow
 
 urlpatterns = [
     path('admin/', admin.site.urls),
